refactor(blorktools): css3d refactor start
