+++ packages/blorktools/src/asset_debugger/modals/html-editor-modal/html-editor-modal.js
@@ -35,8 +35,6 @@ import {
     BINARY_DATA_PROPERTY 
 } from '../../util/glb-utils.js';
 import { updateHtmlIcons } from '../../panels/mesh-panel/mesh-panel';
-import { setCustomDisplay, disableCustomDisplay } from '../../util/custom-animation/css3d-util.js';
-
 // Import Three.js the same way as other files in the codebase
 import * as THREE from 'three';
 import { getIsPreviewActive, setLastTextureUpdateTime } from '../../util/custom-animation/animation-util';
